EXTEND-8 : Add a listener service to manage extendable resource removal. Add an implementation of listener to remove extenders of removed resources.
git-svn-id: http://dev.lutece.paris.fr/svn/lutece/portal/trunk/plugins/extends/module-extend-feedback@48511 bab10101-e421-0410-a517-8ce0973de3ef

+++ src/java/fr/paris/lutece/plugins/extend/modules/feedback/service/extender/FeedbackResourceExtender.java
@@ -115,6 +115,9 @@ public class FeedbackResourceExtender extends AbstractResourceExtender
     @Override
     public void doDeleteResourceAddOn( ResourceExtenderDTO extender )
     {
-        _configService.remove( extender.getIdExtender(  ) );
+        if ( extender.getIdExtender( ) > 0 )
+        {
+            _configService.remove( extender.getIdExtender( ) );
+        }
     }
 }
